# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a k8s management platform API documentation",
        "title": "K8s Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/hosts": {
            "post": {
                "tags": [
                    "主机管理"
                ],
                "summary": "创建主机",
                "parameters": [
                    {
                        "description": "创建主机请求参数",
                        "name": "createHostRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateHostDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/hosts/page": {
            "post": {
                "tags": [
                    "主机管理"
                ],
                "summary": "获取主机列表分页",
                "parameters": [
                    {
                        "description": "主机列表分页请求参数",
                        "name": "hostPageRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HostPageRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/hosts/select": {
            "get": {
                "description": "获取所有主机的下拉选项列表，用于前端选择框展示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "获取主机下拉列表",
                "responses": {
                    "200": {
                        "description": "成功返回主机下拉列表数据",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/hosts/{id}": {
            "put": {
                "tags": [
                    "主机管理"
                ],
                "summary": "更新主机信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新主机请求参数",
                        "name": "updateHostRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateHostDTO"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "主机管理"
                ],
                "summary": "删除主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/hosts/{id}/inspect": {
            "post": {
                "tags": [
                    "主机管理"
                ],
                "summary": "巡检主机状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/hosts/{id}/test": {
            "post": {
                "tags": [
                    "主机管理"
                ],
                "summary": "测试主机连接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/menu": {
            "put": {
                "description": "根据ID更新菜单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "菜单更新信息",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MenuUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的菜单项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "菜单创建信息",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MenuCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/menu/list": {
            "get": {
                "description": "根据条件查询菜单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "菜单状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/menu/options": {
            "get": {
                "description": "获取菜单下拉选项数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单选项",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/menu/tree": {
            "get": {
                "description": "获取完整的菜单树形结构",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/menu/{id}": {
            "get": {
                "description": "根据菜单ID获取菜单详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "根据ID获取菜单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除指定菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "业务处理失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "dto.CreateHostDTO": {
            "type": "object",
            "required": [
                "address",
                "hostName",
                "hostPassword",
                "hostPort",
                "username"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "hostName": {
                    "type": "string"
                },
                "hostPassword": {
                    "type": "string"
                },
                "hostPort": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.HostPageRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
...
        "dto.MenuCreateDTO": {
            "type": "object",
            "required": [
                "name",
                "permission",
                "sort",
                "type"
            ],
            "properties": {
                "always_show": {
                    "type": "boolean"
                },
                "component": {
                    "type": "string"
                },
                "componentName": {
                    "type": "string"
                },
                "icon": {
                    "type": "string",
                    "maxLength": 100
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "parentId": {
                    "type": "integer",
                    "minimum": 0
                },
                "path": {
                    "type": "string",
                    "maxLength": 200
                },
                "permission": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "sort": {
                    "type": "integer",
                    "minimum": 0
                },
                "type": {
                    "description": "1:目录 2:菜单 3:按钮",
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
        "dto.MenuUpdateDTO": {
            "type": "object",
            "required": [
                "id",
                "name",
                "permission",
                "sort",
                "type"
            ],
            "properties": {
                "alwaysShow": {
                    "type": "boolean"
                },
                "component": {
                    "type": "string"
                },
                "componentName": {
                    "type": "string"
                },
                "icon": {
                    "type": "string",
                    "maxLength": 100
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string",
                    "maxLength": 200
                },
                "permission": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "sort": {
                    "type": "integer",
                    "minimum": 0
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateHostDTO": {
            "type": "object",
            "required": [
                "address",
                "hostName",
                "hostPort",
                "id",
                "username"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "hostName": {
                    "type": "string"
                },
                "hostPassword": {
                    "type": "string"
                },
                "hostPort": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}